backend/requirements.txt
autopep8
celery
cryptography
faker
flask
flask_caching
flask_cors
flask_jwt_extended
flask_migrate
flask_sqlalchemy
pipreqs
pymysql
pytest
pytest_cov
pytest_env
pyyaml
redis
